+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Add HTTP/3 flow-control window auto-tuning and get_stream_send_capacity
+
 ### Changed
 
 ### Deprecated